requirements.txt
django
djangorestframework==3.14.0
psycopg2-binary
